my-new-project/requirements.txt
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
tableauserverclient>=0.25
chardet>=5.0.0
